+++ app/Policies/Filament/ActivityPolicy.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace App\Policies;
+namespace App\Policies\Filament;
 
-use App\Models\Pirep;
 use App\Models\User;
 use Illuminate\Auth\Access\HandlesAuthorization;
+use Spatie\Activitylog\Models\Activity;
 
-class PirepPolicy
+class ActivityPolicy
 {
     use HandlesAuthorization;
 
@@ -15,15 +15,15 @@ class PirepPolicy
      */
     public function viewAny(User $user): bool
     {
-        return $user->can('view_any_pirep');
+        return $user->can('view_any_activity::log');
     }
 
     /**
      * Determine whether the user can view the model.
      */
-    public function view(User $user, Pirep $pirep): bool
+    public function view(User $user, Activity $activity): bool
     {
-        return $user->can('view_pirep');
+        return $user->can('view_activity::log');
     }
 
     /**
@@ -31,23 +31,23 @@ class PirepPolicy
      */
     public function create(User $user): bool
     {
-        return $user->can('create_pirep');
+        return $user->can('create_activity::log');
     }
 
     /**
      * Determine whether the user can update the model.
      */
-    public function update(User $user, Pirep $pirep): bool
+    public function update(User $user, Activity $activity): bool
     {
-        return $user->can('update_pirep');
+        return $user->can('update_activity::log');
     }
 
     /**
      * Determine whether the user can delete the model.
      */
-    public function delete(User $user, Pirep $pirep): bool
+    public function delete(User $user, Activity $activity): bool
     {
-        return $user->can('delete_pirep');
+        return $user->can('delete_activity::log');
     }
 
     /**
@@ -55,15 +55,15 @@ class PirepPolicy
      */
     public function deleteAny(User $user): bool
     {
-        return $user->can('delete_any_pirep');
+        return $user->can('delete_any_activity::log');
     }
 
     /**
      * Determine whether the user can permanently delete.
      */
-    public function forceDelete(User $user, Pirep $pirep): bool
+    public function forceDelete(User $user, Activity $activity): bool
     {
-        return $user->can('force_delete_pirep');
+        return $user->can('force_delete_activity::log');
     }
 
     /**
@@ -71,15 +71,15 @@ class PirepPolicy
      */
     public function forceDeleteAny(User $user): bool
     {
-        return $user->can('force_delete_any_pirep');
+        return $user->can('force_delete_any_activity::log');
     }
 
     /**
      * Determine whether the user can restore.
      */
-    public function restore(User $user, Pirep $pirep): bool
+    public function restore(User $user, Activity $activity): bool
     {
-        return $user->can('restore_pirep');
+        return $user->can('restore_activity::log');
     }
 
     /**
@@ -87,15 +87,15 @@ class PirepPolicy
      */
     public function restoreAny(User $user): bool
     {
-        return $user->can('restore_any_pirep');
+        return $user->can('restore_any_activity::log');
     }
 
     /**
      * Determine whether the user can replicate.
      */
-    public function replicate(User $user, Pirep $pirep): bool
+    public function replicate(User $user, Activity $activity): bool
     {
-        return $user->can('replicate_pirep');
+        return $user->can('replicate_activity::log');
     }
 
     /**
@@ -103,6 +103,6 @@ class PirepPolicy
      */
     public function reorder(User $user): bool
     {
-        return $user->can('reorder_pirep');
+        return $user->can('reorder_activity::log');
     }
 }
