+++ app/Policies/Filament/AirlinePolicy.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace App\Policies;
+namespace App\Policies\Filament;
 
-use App\Models\Typerating;
+use App\Models\Airline;
 use App\Models\User;
 use Illuminate\Auth\Access\HandlesAuthorization;
 
-class TyperatingPolicy
+class AirlinePolicy
 {
     use HandlesAuthorization;
 
@@ -15,15 +15,15 @@ class TyperatingPolicy
      */
     public function viewAny(User $user): bool
     {
-        return $user->can('view_any_typerating');
+        return $user->can('view_any_airline');
     }
 
     /**
      * Determine whether the user can view the model.
      */
-    public function view(User $user, Typerating $typerating): bool
+    public function view(User $user, Airline $airline): bool
     {
-        return $user->can('view_typerating');
+        return $user->can('view_airline');
     }
 
     /**
@@ -31,23 +31,23 @@ class TyperatingPolicy
      */
     public function create(User $user): bool
     {
-        return $user->can('create_typerating');
+        return $user->can('create_airline');
     }
 
     /**
      * Determine whether the user can update the model.
      */
-    public function update(User $user, Typerating $typerating): bool
+    public function update(User $user, Airline $airline): bool
     {
-        return $user->can('update_typerating');
+        return $user->can('update_airline');
     }
 
     /**
      * Determine whether the user can delete the model.
      */
-    public function delete(User $user, Typerating $typerating): bool
+    public function delete(User $user, Airline $airline): bool
     {
-        return $user->can('delete_typerating');
+        return $user->can('delete_airline');
     }
 
     /**
@@ -55,15 +55,15 @@ class TyperatingPolicy
      */
     public function deleteAny(User $user): bool
     {
-        return $user->can('delete_any_typerating');
+        return $user->can('delete_any_airline');
     }
 
     /**
      * Determine whether the user can permanently delete.
      */
-    public function forceDelete(User $user, Typerating $typerating): bool
+    public function forceDelete(User $user, Airline $airline): bool
     {
-        return $user->can('force_delete_typerating');
+        return $user->can('force_delete_airline');
     }
 
     /**
@@ -71,15 +71,15 @@ class TyperatingPolicy
      */
     public function forceDeleteAny(User $user): bool
     {
-        return $user->can('force_delete_any_typerating');
+        return $user->can('force_delete_any_airline');
     }
 
     /**
      * Determine whether the user can restore.
      */
-    public function restore(User $user, Typerating $typerating): bool
+    public function restore(User $user, Airline $airline): bool
     {
-        return $user->can('restore_typerating');
+        return $user->can('restore_airline');
     }
 
     /**
@@ -87,15 +87,15 @@ class TyperatingPolicy
      */
     public function restoreAny(User $user): bool
     {
-        return $user->can('restore_any_typerating');
+        return $user->can('restore_any_airline');
     }
 
     /**
      * Determine whether the user can replicate.
      */
-    public function replicate(User $user, Typerating $typerating): bool
+    public function replicate(User $user, Airline $airline): bool
     {
-        return $user->can('replicate_typerating');
+        return $user->can('replicate_airline');
     }
 
     /**
@@ -103,6 +103,6 @@ class TyperatingPolicy
      */
     public function reorder(User $user): bool
     {
-        return $user->can('reorder_typerating');
+        return $user->can('reorder_airline');
     }
 }
