+++ app/Policies/Filament/AirportPolicy.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace App\Policies;
+namespace App\Policies\Filament;
 
-use App\Models\SimBriefAirframe;
+use App\Models\Airport;
 use App\Models\User;
 use Illuminate\Auth\Access\HandlesAuthorization;
 
-class SimBriefAirframePolicy
+class AirportPolicy
 {
     use HandlesAuthorization;
 
@@ -15,15 +15,15 @@ class SimBriefAirframePolicy
      */
     public function viewAny(User $user): bool
     {
-        return $user->can('view_any_airframe');
+        return $user->can('view_any_airport');
     }
 
     /**
      * Determine whether the user can view the model.
      */
-    public function view(User $user, SimBriefAirframe $simBriefAirframe): bool
+    public function view(User $user, Airport $airport): bool
     {
-        return $user->can('view_airframe');
+        return $user->can('view_airport');
     }
 
     /**
@@ -31,23 +31,23 @@ class SimBriefAirframePolicy
      */
     public function create(User $user): bool
     {
-        return $user->can('create_airframe');
+        return $user->can('create_airport');
     }
 
     /**
      * Determine whether the user can update the model.
      */
-    public function update(User $user, SimBriefAirframe $simBriefAirframe): bool
+    public function update(User $user, Airport $airport): bool
     {
-        return $user->can('update_airframe');
+        return $user->can('update_airport');
     }
 
     /**
      * Determine whether the user can delete the model.
      */
-    public function delete(User $user, SimBriefAirframe $simBriefAirframe): bool
+    public function delete(User $user, Airport $airport): bool
     {
-        return $user->can('delete_airframe');
+        return $user->can('delete_airport');
     }
 
     /**
@@ -55,15 +55,15 @@ class SimBriefAirframePolicy
      */
     public function deleteAny(User $user): bool
     {
-        return $user->can('delete_any_airframe');
+        return $user->can('delete_any_airport');
     }
 
     /**
      * Determine whether the user can permanently delete.
      */
-    public function forceDelete(User $user, SimBriefAirframe $simBriefAirframe): bool
+    public function forceDelete(User $user, Airport $airport): bool
     {
-        return $user->can('force_delete_airframe');
+        return $user->can('force_delete_airport');
     }
 
     /**
@@ -71,15 +71,15 @@ class SimBriefAirframePolicy
      */
     public function forceDeleteAny(User $user): bool
     {
-        return $user->can('force_delete_any_airframe');
+        return $user->can('force_delete_any_airport');
     }
 
     /**
      * Determine whether the user can restore.
      */
-    public function restore(User $user, SimBriefAirframe $simBriefAirframe): bool
+    public function restore(User $user, Airport $airport): bool
     {
-        return $user->can('restore_airframe');
+        return $user->can('restore_airport');
     }
 
     /**
@@ -87,15 +87,15 @@ class SimBriefAirframePolicy
      */
     public function restoreAny(User $user): bool
     {
-        return $user->can('restore_any_airframe');
+        return $user->can('restore_any_airport');
     }
 
     /**
      * Determine whether the user can replicate.
      */
-    public function replicate(User $user, SimBriefAirframe $simBriefAirframe): bool
+    public function replicate(User $user, Airport $airport): bool
     {
-        return $user->can('replicate_airframe');
+        return $user->can('replicate_airport');
     }
 
     /**
@@ -103,6 +103,6 @@ class SimBriefAirframePolicy
      */
     public function reorder(User $user): bool
     {
-        return $user->can('reorder_airframe');
+        return $user->can('reorder_airport');
     }
 }
