+++ app/Policies/Filament/ExpensePolicy.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace App\Policies;
+namespace App\Policies\Filament;
 
-use App\Models\Award;
+use App\Models\Expense;
 use App\Models\User;
 use Illuminate\Auth\Access\HandlesAuthorization;
 
-class AwardPolicy
+class ExpensePolicy
 {
     use HandlesAuthorization;
 
@@ -15,15 +15,15 @@ class AwardPolicy
      */
     public function viewAny(User $user): bool
     {
-        return $user->can('view_any_award');
+        return $user->can('view_any_expense');
     }
 
     /**
      * Determine whether the user can view the model.
      */
-    public function view(User $user, Award $award): bool
+    public function view(User $user, Expense $expense): bool
     {
-        return $user->can('view_award');
+        return $user->can('view_expense');
     }
 
     /**
@@ -31,23 +31,23 @@ class AwardPolicy
      */
     public function create(User $user): bool
     {
-        return $user->can('create_award');
+        return $user->can('create_expense');
     }
 
     /**
      * Determine whether the user can update the model.
      */
-    public function update(User $user, Award $award): bool
+    public function update(User $user, Expense $expense): bool
     {
-        return $user->can('update_award');
+        return $user->can('update_expense');
     }
 
     /**
      * Determine whether the user can delete the model.
      */
-    public function delete(User $user, Award $award): bool
+    public function delete(User $user, Expense $expense): bool
     {
-        return $user->can('delete_award');
+        return $user->can('delete_expense');
     }
 
     /**
@@ -55,15 +55,15 @@ class AwardPolicy
      */
     public function deleteAny(User $user): bool
     {
-        return $user->can('delete_any_award');
+        return $user->can('delete_any_expense');
     }
 
     /**
      * Determine whether the user can permanently delete.
      */
-    public function forceDelete(User $user, Award $award): bool
+    public function forceDelete(User $user, Expense $expense): bool
     {
-        return $user->can('force_delete_award');
+        return $user->can('force_delete_expense');
     }
 
     /**
@@ -71,15 +71,15 @@ class AwardPolicy
      */
     public function forceDeleteAny(User $user): bool
     {
-        return $user->can('force_delete_any_award');
+        return $user->can('force_delete_any_expense');
     }
 
     /**
      * Determine whether the user can restore.
      */
-    public function restore(User $user, Award $award): bool
+    public function restore(User $user, Expense $expense): bool
     {
-        return $user->can('restore_award');
+        return $user->can('restore_expense');
     }
 
     /**
@@ -87,15 +87,15 @@ class AwardPolicy
      */
     public function restoreAny(User $user): bool
     {
-        return $user->can('restore_any_award');
+        return $user->can('restore_any_expense');
     }
 
     /**
      * Determine whether the user can replicate.
      */
-    public function replicate(User $user, Award $award): bool
+    public function replicate(User $user, Expense $expense): bool
     {
-        return $user->can('replicate_award');
+        return $user->can('replicate_expense');
     }
 
     /**
@@ -103,6 +103,6 @@ class AwardPolicy
      */
     public function reorder(User $user): bool
     {
-        return $user->can('reorder_award');
+        return $user->can('reorder_expense');
     }
 }
