+++ app/Policies/Filament/FlightPolicy.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace App\Policies;
+namespace App\Policies\Filament;
 
-use App\Models\PirepField;
+use App\Models\Flight;
 use App\Models\User;
 use Illuminate\Auth\Access\HandlesAuthorization;
 
-class PirepFieldPolicy
+class FlightPolicy
 {
     use HandlesAuthorization;
 
@@ -15,15 +15,15 @@ class PirepFieldPolicy
      */
     public function viewAny(User $user): bool
     {
-        return $user->can('view_any_pirep::field');
+        return $user->can('view_any_flight');
     }
 
     /**
      * Determine whether the user can view the model.
      */
-    public function view(User $user, PirepField $pirepField): bool
+    public function view(User $user, Flight $flight): bool
     {
-        return $user->can('view_pirep::field');
+        return $user->can('view_flight');
     }
 
     /**
@@ -31,23 +31,23 @@ class PirepFieldPolicy
      */
     public function create(User $user): bool
     {
-        return $user->can('create_pirep::field');
+        return $user->can('create_flight');
     }
 
     /**
      * Determine whether the user can update the model.
      */
-    public function update(User $user, PirepField $pirepField): bool
+    public function update(User $user, Flight $flight): bool
     {
-        return $user->can('update_pirep::field');
+        return $user->can('update_flight');
     }
 
     /**
      * Determine whether the user can delete the model.
      */
-    public function delete(User $user, PirepField $pirepField): bool
+    public function delete(User $user, Flight $flight): bool
     {
-        return $user->can('delete_pirep::field');
+        return $user->can('delete_flight');
     }
 
     /**
@@ -55,15 +55,15 @@ class PirepFieldPolicy
      */
     public function deleteAny(User $user): bool
     {
-        return $user->can('delete_any_pirep::field');
+        return $user->can('delete_any_flight');
     }
 
     /**
      * Determine whether the user can permanently delete.
      */
-    public function forceDelete(User $user, PirepField $pirepField): bool
+    public function forceDelete(User $user, Flight $flight): bool
     {
-        return $user->can('force_delete_pirep::field');
+        return $user->can('force_delete_flight');
     }
 
     /**
@@ -71,15 +71,15 @@ class PirepFieldPolicy
      */
     public function forceDeleteAny(User $user): bool
     {
-        return $user->can('force_delete_any_pirep::field');
+        return $user->can('force_delete_any_flight');
     }
 
     /**
      * Determine whether the user can restore.
      */
-    public function restore(User $user, PirepField $pirepField): bool
+    public function restore(User $user, Flight $flight): bool
     {
-        return $user->can('restore_pirep::field');
+        return $user->can('restore_flight');
     }
 
     /**
@@ -87,15 +87,15 @@ class PirepFieldPolicy
      */
     public function restoreAny(User $user): bool
     {
-        return $user->can('restore_any_pirep::field');
+        return $user->can('restore_any_flight');
     }
 
     /**
      * Determine whether the user can replicate.
      */
-    public function replicate(User $user, PirepField $pirepField): bool
+    public function replicate(User $user, Flight $flight): bool
     {
-        return $user->can('replicate_pirep::field');
+        return $user->can('replicate_flight');
     }
 
     /**
@@ -103,6 +103,6 @@ class PirepFieldPolicy
      */
     public function reorder(User $user): bool
     {
-        return $user->can('reorder_pirep::field');
+        return $user->can('reorder_flight');
     }
 }
