+++ app/Policies/Filament/PirepFieldPolicy.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace App\Policies;
+namespace App\Policies\Filament;
 
+use App\Models\PirepField;
 use App\Models\User;
 use Illuminate\Auth\Access\HandlesAuthorization;
-use Spatie\Activitylog\Models\Activity;
 
-class ActivityPolicy
+class PirepFieldPolicy
 {
     use HandlesAuthorization;
 
@@ -15,15 +15,15 @@ class ActivityPolicy
      */
     public function viewAny(User $user): bool
     {
-        return $user->can('view_any_activity::log');
+        return $user->can('view_any_pirep::field');
     }
 
     /**
      * Determine whether the user can view the model.
      */
-    public function view(User $user, Activity $activity): bool
+    public function view(User $user, PirepField $pirepField): bool
     {
-        return $user->can('view_activity::log');
+        return $user->can('view_pirep::field');
     }
 
     /**
@@ -31,23 +31,23 @@ class ActivityPolicy
      */
     public function create(User $user): bool
     {
-        return $user->can('create_activity::log');
+        return $user->can('create_pirep::field');
     }
 
     /**
      * Determine whether the user can update the model.
      */
-    public function update(User $user, Activity $activity): bool
+    public function update(User $user, PirepField $pirepField): bool
     {
-        return $user->can('update_activity::log');
+        return $user->can('update_pirep::field');
     }
 
     /**
      * Determine whether the user can delete the model.
      */
-    public function delete(User $user, Activity $activity): bool
+    public function delete(User $user, PirepField $pirepField): bool
     {
-        return $user->can('delete_activity::log');
+        return $user->can('delete_pirep::field');
     }
 
     /**
@@ -55,15 +55,15 @@ class ActivityPolicy
      */
     public function deleteAny(User $user): bool
     {
-        return $user->can('delete_any_activity::log');
+        return $user->can('delete_any_pirep::field');
     }
 
     /**
      * Determine whether the user can permanently delete.
      */
-    public function forceDelete(User $user, Activity $activity): bool
+    public function forceDelete(User $user, PirepField $pirepField): bool
     {
-        return $user->can('force_delete_activity::log');
+        return $user->can('force_delete_pirep::field');
     }
 
     /**
@@ -71,15 +71,15 @@ class ActivityPolicy
      */
     public function forceDeleteAny(User $user): bool
     {
-        return $user->can('force_delete_any_activity::log');
+        return $user->can('force_delete_any_pirep::field');
     }
 
     /**
      * Determine whether the user can restore.
      */
-    public function restore(User $user, Activity $activity): bool
+    public function restore(User $user, PirepField $pirepField): bool
     {
-        return $user->can('restore_activity::log');
+        return $user->can('restore_pirep::field');
     }
 
     /**
@@ -87,15 +87,15 @@ class ActivityPolicy
      */
     public function restoreAny(User $user): bool
     {
-        return $user->can('restore_any_activity::log');
+        return $user->can('restore_any_pirep::field');
     }
 
     /**
      * Determine whether the user can replicate.
      */
-    public function replicate(User $user, Activity $activity): bool
+    public function replicate(User $user, PirepField $pirepField): bool
     {
-        return $user->can('replicate_activity::log');
+        return $user->can('replicate_pirep::field');
     }
 
     /**
@@ -103,6 +103,6 @@ class ActivityPolicy
      */
     public function reorder(User $user): bool
     {
-        return $user->can('reorder_activity::log');
+        return $user->can('reorder_pirep::field');
     }
 }
