+++ app/Policies/Filament/RankPolicy.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace App\Policies;
+namespace App\Policies\Filament;
 
-use App\Models\Subfleet;
+use App\Models\Rank;
 use App\Models\User;
 use Illuminate\Auth\Access\HandlesAuthorization;
 
-class SubfleetPolicy
+class RankPolicy
 {
     use HandlesAuthorization;
 
@@ -15,15 +15,15 @@ class SubfleetPolicy
      */
     public function viewAny(User $user): bool
     {
-        return $user->can('view_any_subfleet');
+        return $user->can('view_any_rank');
     }
 
     /**
      * Determine whether the user can view the model.
      */
-    public function view(User $user, Subfleet $subfleet): bool
+    public function view(User $user, Rank $rank): bool
     {
-        return $user->can('view_subfleet');
+        return $user->can('view_rank');
     }
 
     /**
@@ -31,23 +31,23 @@ class SubfleetPolicy
      */
     public function create(User $user): bool
     {
-        return $user->can('create_subfleet');
+        return $user->can('create_rank');
     }
 
     /**
      * Determine whether the user can update the model.
      */
-    public function update(User $user, Subfleet $subfleet): bool
+    public function update(User $user, Rank $rank): bool
     {
-        return $user->can('update_subfleet');
+        return $user->can('update_rank');
     }
 
     /**
      * Determine whether the user can delete the model.
      */
-    public function delete(User $user, Subfleet $subfleet): bool
+    public function delete(User $user, Rank $rank): bool
     {
-        return $user->can('delete_subfleet');
+        return $user->can('delete_rank');
     }
 
     /**
@@ -55,15 +55,15 @@ class SubfleetPolicy
      */
     public function deleteAny(User $user): bool
     {
-        return $user->can('delete_any_subfleet');
+        return $user->can('delete_any_rank');
     }
 
     /**
      * Determine whether the user can permanently delete.
      */
-    public function forceDelete(User $user, Subfleet $subfleet): bool
+    public function forceDelete(User $user, Rank $rank): bool
     {
-        return $user->can('force_delete_subfleet');
+        return $user->can('force_delete_rank');
     }
 
     /**
@@ -71,15 +71,15 @@ class SubfleetPolicy
      */
     public function forceDeleteAny(User $user): bool
     {
-        return $user->can('force_delete_any_subfleet');
+        return $user->can('force_delete_any_rank');
     }
 
     /**
      * Determine whether the user can restore.
      */
-    public function restore(User $user, Subfleet $subfleet): bool
+    public function restore(User $user, Rank $rank): bool
     {
-        return $user->can('restore_subfleet');
+        return $user->can('restore_rank');
     }
 
     /**
@@ -87,15 +87,15 @@ class SubfleetPolicy
      */
     public function restoreAny(User $user): bool
     {
-        return $user->can('restore_any_subfleet');
+        return $user->can('restore_any_rank');
     }
 
     /**
      * Determine whether the user can replicate.
      */
-    public function replicate(User $user, Subfleet $subfleet): bool
+    public function replicate(User $user, Rank $rank): bool
     {
-        return $user->can('replicate_subfleet');
+        return $user->can('replicate_rank');
     }
 
     /**
@@ -103,6 +103,6 @@ class SubfleetPolicy
      */
     public function reorder(User $user): bool
     {
-        return $user->can('reorder_subfleet');
+        return $user->can('reorder_rank');
     }
 }
