+++ app/Policies/Filament/SubfleetPolicy.php
@@ -1,11 +1,12 @@
 <?php
 
-namespace App\Policies;
+namespace App\Policies\Filament;
 
+use App\Models\Subfleet;
 use App\Models\User;
 use Illuminate\Auth\Access\HandlesAuthorization;
 
-class UserPolicy
+class SubfleetPolicy
 {
     use HandlesAuthorization;
 
@@ -14,15 +15,15 @@ class UserPolicy
      */
     public function viewAny(User $user): bool
     {
-        return $user->can('view_any_user');
+        return $user->can('view_any_subfleet');
     }
 
     /**
      * Determine whether the user can view the model.
      */
-    public function view(User $user): bool
+    public function view(User $user, Subfleet $subfleet): bool
     {
-        return $user->can('view_user');
+        return $user->can('view_subfleet');
     }
 
     /**
@@ -30,23 +31,23 @@ class UserPolicy
      */
     public function create(User $user): bool
     {
-        return $user->can('create_user');
+        return $user->can('create_subfleet');
     }
 
     /**
      * Determine whether the user can update the model.
      */
-    public function update(User $user): bool
+    public function update(User $user, Subfleet $subfleet): bool
     {
-        return $user->can('update_user');
+        return $user->can('update_subfleet');
     }
 
     /**
      * Determine whether the user can delete the model.
      */
-    public function delete(User $user): bool
+    public function delete(User $user, Subfleet $subfleet): bool
     {
-        return $user->can('delete_user');
+        return $user->can('delete_subfleet');
     }
 
     /**
@@ -54,15 +55,15 @@ class UserPolicy
      */
     public function deleteAny(User $user): bool
     {
-        return $user->can('delete_any_user');
+        return $user->can('delete_any_subfleet');
     }
 
     /**
      * Determine whether the user can permanently delete.
      */
-    public function forceDelete(User $user): bool
+    public function forceDelete(User $user, Subfleet $subfleet): bool
     {
-        return $user->can('force_delete_user');
+        return $user->can('force_delete_subfleet');
     }
 
     /**
@@ -70,15 +71,15 @@ class UserPolicy
      */
     public function forceDeleteAny(User $user): bool
     {
-        return $user->can('force_delete_any_user');
+        return $user->can('force_delete_any_subfleet');
     }
 
     /**
      * Determine whether the user can restore.
      */
-    public function restore(User $user): bool
+    public function restore(User $user, Subfleet $subfleet): bool
     {
-        return $user->can('restore_user');
+        return $user->can('restore_subfleet');
     }
 
     /**
@@ -86,15 +87,15 @@ class UserPolicy
      */
     public function restoreAny(User $user): bool
     {
-        return $user->can('restore_any_user');
+        return $user->can('restore_any_subfleet');
     }
 
     /**
-     * Determine whether the user can bulk restore.
+     * Determine whether the user can replicate.
      */
-    public function replicate(User $user): bool
+    public function replicate(User $user, Subfleet $subfleet): bool
     {
-        return $user->can('replicate_user');
+        return $user->can('replicate_subfleet');
     }
 
     /**
@@ -102,6 +103,6 @@ class UserPolicy
      */
     public function reorder(User $user): bool
     {
-        return $user->can('reorder_user');
+        return $user->can('reorder_subfleet');
     }
 }
