+++ app/Policies/Filament/UserPolicy.php
@@ -1,12 +1,11 @@
 <?php
 
-namespace App\Policies;
+namespace App\Policies\Filament;
 
-use App\Models\Aircraft;
 use App\Models\User;
 use Illuminate\Auth\Access\HandlesAuthorization;
 
-class AircraftPolicy
+class UserPolicy
 {
     use HandlesAuthorization;
 
@@ -15,15 +14,15 @@ class AircraftPolicy
      */
     public function viewAny(User $user): bool
     {
-        return $user->can('view_any_aircraft');
+        return $user->can('view_any_user');
     }
 
     /**
      * Determine whether the user can view the model.
      */
-    public function view(User $user, Aircraft $aircraft): bool
+    public function view(User $user): bool
     {
-        return $user->can('view_aircraft');
+        return $user->can('view_user');
     }
 
     /**
@@ -31,23 +30,23 @@ class AircraftPolicy
      */
     public function create(User $user): bool
     {
-        return $user->can('create_aircraft');
+        return $user->can('create_user');
     }
 
     /**
      * Determine whether the user can update the model.
      */
-    public function update(User $user, Aircraft $aircraft): bool
+    public function update(User $user): bool
     {
-        return $user->can('update_aircraft');
+        return $user->can('update_user');
     }
 
     /**
      * Determine whether the user can delete the model.
      */
-    public function delete(User $user, Aircraft $aircraft): bool
+    public function delete(User $user): bool
     {
-        return $user->can('delete_aircraft');
+        return $user->can('delete_user');
     }
 
     /**
@@ -55,15 +54,15 @@ class AircraftPolicy
      */
     public function deleteAny(User $user): bool
     {
-        return $user->can('delete_any_aircraft');
+        return $user->can('delete_any_user');
     }
 
     /**
      * Determine whether the user can permanently delete.
      */
-    public function forceDelete(User $user, Aircraft $aircraft): bool
+    public function forceDelete(User $user): bool
     {
-        return $user->can('force_delete_aircraft');
+        return $user->can('force_delete_user');
     }
 
     /**
@@ -71,15 +70,15 @@ class AircraftPolicy
      */
     public function forceDeleteAny(User $user): bool
     {
-        return $user->can('force_delete_any_aircraft');
+        return $user->can('force_delete_any_user');
     }
 
     /**
      * Determine whether the user can restore.
      */
-    public function restore(User $user, Aircraft $aircraft): bool
+    public function restore(User $user): bool
     {
-        return $user->can('restore_aircraft');
+        return $user->can('restore_user');
     }
 
     /**
@@ -87,15 +86,15 @@ class AircraftPolicy
      */
     public function restoreAny(User $user): bool
     {
-        return $user->can('restore_any_aircraft');
+        return $user->can('restore_any_user');
     }
 
     /**
-     * Determine whether the user can replicate.
+     * Determine whether the user can bulk restore.
      */
-    public function replicate(User $user, Aircraft $aircraft): bool
+    public function replicate(User $user): bool
     {
-        return $user->can('replicate_aircraft');
+        return $user->can('replicate_user');
     }
 
     /**
@@ -103,6 +102,6 @@ class AircraftPolicy
      */
     public function reorder(User $user): bool
     {
-        return $user->can('reorder_aircraft');
+        return $user->can('reorder_user');
     }
 }
